backend/requirements.txt
# Core Django
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0

# Database
psycopg2-binary==2.9.11
django-db-connection-pool==1.2.6

# Environment & Configuration
django-environ==0.12.0

# CORS & Security
django-cors-headers==4.9.0

# ASGI Server
uvicorn[standard]
uvloop

# Image Processing
Pillow==12.0.0
# pillow-simd==12.0.0  # Optional: Faster image processing (version 12.0.0 not available, using Pillow instead)

celery==5.5.3
redis>=5.0.0,<6.0.0  # Compatible with django-redis 5.4.0 (redis 7.x has breaking changes)

# Rich Text Editor
django-ckeditor-5==0.2.18

# Filtering & Search
django-filter==25.2

# Analytics & Tracking
django-hitcount==1.3.5

# Spam Protection
django-recaptcha==4.1.0

# Backups
django-dbbackup==5.0.1

# AI Chatbot
groq==0.34.0

# HTTP Requests
requests==2.32.5

# Production Server
gunicorn==21.2.0

# Core Dependencies (automatically installed)
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
urllib3==2.5.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11

# Celery Dependencies
amqp==5.3.1
billiard==4.2.2
kombu==5.5.4
vine==5.1.0
click==8.3.0

# HTTP Client (for GROQ)
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
sniffio==1.3.1
anyio==4.11.0

# JSON Schema (for drf-spectacular)
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.28.0
uritemplate==4.2.0

# Pydantic (for groq)
pydantic==2.12.4
pydantic_core==2.41.5
annotated-types==0.7.0
typing_extensions==4.15.0

# Utilities
attrs==25.4.0
packaging>=23.2,<25
six==1.17.0
inflection==0.5.1
distro==1.9.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
prompt_toolkit==3.0.52
wcwidth==0.2.14
typing-inspection==0.4.2

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-mock==3.14.0
factory-boy==3.3.1
faker==33.1.0

# Caching
django-redis==5.4.0

# Code Quality
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.11.2

# Monitoring & Observability
sentry-sdk==2.19.0
python-json-logger==2.0.7
psutil==6.1.0


# PDF Generation
reportlab==4.2.5
pypdf==5.1.0
